@@ -3,7 +3,7 @@ description: Demonstrates how to use the google_map_polyutil plugin.
 publish_to: "none"
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.16.2 <3.0.0"
 
 dependencies:
   flutter:
